batch: Rename lifecycle module
Batch create, delete, and note updates are lifecycle operations for the
durable batch object. Naming that module operations made it hard to
distinguish from other generic operation modules such as
staging.operations.

The module name also hid the fact that callers are touching batch lifecycle
state rather than arbitrary batch mechanics. A more specific name makes the
batch package easier to scan.

This commit addresses that by renaming batch.operations to batch.lifecycle
and updating production imports to use the new module name without leaving
an old compatibility module behind.

Follow-up test commits will move import sites in tests and guard the module
name boundary.
