+++ src/git_stage_batch/batch/lifecycle.py
