+++ src/git_stage_batch/batch/text_file_storage.py
@@ -7,7 +7,7 @@ from dataclasses import dataclass
 from typing import TYPE_CHECKING
 
 from .metadata_validation import get_validated_baseline_commit
-from .operations import create_batch
+from .lifecycle import create_batch
 from .query import get_batch_commit_sha, read_batch_metadata
 from .validation import batch_exists, validate_batch_name
 from ..core.text_lifecycle import (
