+++ src/git_stage_batch/commands/batch_source/reset_claims.py
@@ -7,7 +7,7 @@ import shlex
 from collections.abc import Sequence
 from contextlib import AbstractContextManager
 
-from ...batch.operations import create_batch
+from ...batch.lifecycle import create_batch
 from ...batch.ownership import BatchOwnership
 from ...batch.ownership_detachment import acquire_detached_batch_ownership
 from ...batch.ownership_merging import merge_batch_ownership
