+++ src/git_stage_batch/commands/file_scope/discard_to_batch.py
@@ -7,7 +7,7 @@ from contextlib import ExitStack
 from dataclasses import dataclass
 
 from ...batch.display import annotate_with_batch_source
-from ...batch.operations import create_batch
+from ...batch.lifecycle import create_batch
 from ...batch.query import read_batch_metadata
 from ...batch.source_refresh import acquire_batch_ownership_update_for_selection
 from ...batch.text_file_storage import BatchFileUpdate, add_files_to_batch
