+++ src/git_stage_batch/commands/new.py
@@ -2,7 +2,7 @@
 
 from __future__ import annotations
 
-from ..batch.operations import create_batch
+from ..batch.lifecycle import create_batch
 from ..batch.source_selector import require_plain_batch_name
 import sys
 from ..data.undo import undo_checkpoint
