+++ src/git_stage_batch/commands/selection/selected_change_batch_discarding.py
@@ -7,7 +7,7 @@ from contextlib import ExitStack
 import sys
 
 from ...batch.display import annotate_with_batch_source
-from ...batch.operations import create_batch
+from ...batch.lifecycle import create_batch
 from ...batch.query import read_batch_metadata
 from ...batch.source_refresh import acquire_batch_ownership_update_for_selection
 from ...batch.text_file_storage import add_file_to_batch
