+++ src/git_stage_batch/commands/sift.py
@@ -23,7 +23,7 @@ import sys
 
 from ..exceptions import MergeError
 from ..batch.metadata_validation import read_validated_batch_metadata
-from ..batch.operations import create_batch, delete_batch
+from ..batch.lifecycle import create_batch, delete_batch
 from ..batch.query import read_batch_metadata
 from ..batch.validation import batch_exists, validate_batch_name
 from ..batch.source_selector import require_plain_batch_name
